+++ android/src/main/kotlin/com/mapbox/maps/mapbox_maps/pigeons/GestureListeners.kt
@@ -167,4 +167,20 @@ class GestureListener(private val binaryMessenger: BinaryMessenger, private val 
       }
     }
   }
+  fun onZoom(contextArg: MapContentGestureContext, callback: (Result<Unit>) -> Unit) {
+    val separatedMessageChannelSuffix = if (messageChannelSuffix.isNotEmpty()) ".$messageChannelSuffix" else ""
+    val channelName = "dev.flutter.pigeon.mapbox_maps_flutter.GestureListener.onZoom$separatedMessageChannelSuffix"
+    val channel = BasicMessageChannel<Any?>(binaryMessenger, channelName, codec)
+    channel.send(listOf(contextArg)) {
+      if (it is List<*>) {
+        if (it.size > 1) {
+          callback(Result.failure(FlutterError(it[0] as String, it[1] as String, it[2] as String?)))
+        } else {
+          callback(Result.success(Unit))
+        }
+      } else {
+        callback(Result.failure(createConnectionError(channelName)))
+      }
+    }
+  }
 }